+++ A32NX/html_ui/Pages/VCockpit/Instruments/Airliners/A320_Neo/CDU/CFDS/TEST_INST/A320_Neo_CDU_CFDS_Test_Inst.js
@@ -14,7 +14,7 @@ class CDUCfdsTestInst {
             [""],
             [""],
             [""],
-            ["<RETURN[color]blue"]
+            ["<RETURN[color]cyan"]
         ]);
 
         mcdu.leftInputDelay[0] = () => {