+++ .github/CHANGELOG.md
@@ -110,6 +110,7 @@
 1. [ECAM] Implemented proper FWC flight phases - @beheh (Benedict Etzel)
 1. [ECAM] Updated ECAM messages - @beheh (Benedict Etzel)
 1. [CDU] New improved RADNAV page | Ability to tune navaids with identifier - @St54Kevin (Kevin Karas)
+1. [CDU] Corrected visuals for constraints on flight plan page - @MisterChocker (Leon)
 1. [ND] Add heading line to ARC and NAV map mode @ilyeshammadi (Ilyes Hammadi)
 1. [PFD] Add QNH flashing when reach transition altitude - @Kimbyeoungjang (김병장#7165)
 1. [ECAM] Added flight phase inhibit override to recall button - @beheh (Benedict Etzel)